+++ TasteTrailExperience/src/TasteTrailExperience.Infrastructure/Venues/Services/VenueService.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using TasteTrailData.Core.Users.Models;
 using TasteTrailData.Core.Venues.Models;
+using TasteTrailExperience.Core.Common.Exceptions;
 using TasteTrailExperience.Core.Venues.Dtos;
 using TasteTrailExperience.Core.Venues.Repositories;
 using TasteTrailExperience.Core.Venues.Services;
@@ -17,9 +18,12 @@ public class VenueService : IVenueService
         _venueRepository = venueRepository;
     }
 
-    public async Task<List<VenueGetByCountDto>> GetVenuesByCountAsync(int count)
+    public async Task<List<VenueGetByCountDto>> GetVenuesFromToAsync(int from, int to)
     {
-        var venues = await _venueRepository.GetByCountAsync(count);
+        if (from <= 0 || to <= 0 || from > to)
+            throw new ArgumentException("Invalid 'from' and/or 'to' values.");
+
+        var venues = await _venueRepository.GetFromToAsync(from, to);
 
         var venueDtos = venues.Select(venue => new VenueGetByCountDto
         {
@@ -29,6 +33,7 @@ public class VenueService : IVenueService
             Description = venue.Description,
             Email = venue.Email,
             ContactNumber = venue.ContactNumber,
+            LogoUrlPath = venue.LogoUrlPath,
             AveragePrice = venue.AveragePrice,
             OverallRating = venue.OverallRating,
             UserId = venue.UserId
@@ -62,8 +67,16 @@ public class VenueService : IVenueService
         return venueId;
     }
 
-    public async Task<int?> DeleteVenueByIdAsync(int id)
+    public async Task<int?> DeleteVenueByIdAsync(int id, User user)
     {
+        var venue = await _venueRepository.GetAsNoTrackingAsync(id);
+
+        if (venue is null)
+            return null;
+
+        if (venue.UserId != user.Id) 
+            throw new ForbiddenAccessException();
+
         var venueId = await _venueRepository.DeleteByIdAsync(id);
 
         return venueId;
@@ -71,6 +84,14 @@ public class VenueService : IVenueService
 
     public async Task<int?> PutVenueAsync(VenueUpdateDto venue, User user)
     {
+        var venueToUpdate = await _venueRepository.GetAsNoTrackingAsync(venue.Id);
+
+        if (venueToUpdate is null)
+            return null;
+
+        if (venueToUpdate.UserId != user.Id)
+            throw new ForbiddenAccessException();
+
         var updatedVenue = new Venue() {
             Id = venue.Id,
             Name = venue.Name,
